added more to products table

+++ Bangazon_Terminal_Interface/Bangazon_Terminal_Interface/Bangazon/IProductRepository.cs
@@ -13,7 +13,7 @@ namespace Bangazon_Terminal_Interface.Bangazon
 
         void AddProductId(int productId);
 
-        void AddProduct(string productName, int productPrice);
+        void AddProduct(string productName, double productPrice);
         
     }
 }
